Refactor Actor into an abstract class and move common fields up.
- Converted `Actor` interface to an `abstract class`.
- Moved `uuid`, `particlePool`, `tile`, `tileCoord`, `health`, `dead`, `status`, `inventory`, and `name` fields to `Actor`.
- Annotated `Actor` with `@Derializable`.
- Updated `CardPlayer` and `Structure` to extend `Actor` and remove redundant fields.
- Centralized common `reindex` logic in `Actor`.
- Updated `ProcChainTest` to adapt to the new `Actor` structure.
