+++ nomad-realms/src/main/java/nomadrealms/context/game/actor/Actor.java
@@ -4,6 +4,7 @@ import static engine.visuals.constraint.misc.TimedConstraint.time;
 import static engine.visuals.constraint.posdim.AbsoluteConstraint.absolute;
 import static nomadrealms.context.game.actor.status.StatusEffect.INVINCIBLE;
 
+import engine.serialization.Derializable;
 import engine.visuals.constraint.box.ConstraintPair;
 import java.util.ArrayList;
 import java.util.List;
@@ -18,10 +19,13 @@ import nomadrealms.context.game.card.effect.SpawnParticlesEffect;
 import nomadrealms.context.game.card.query.actor.SelfQuery;
 import nomadrealms.context.game.event.InputEvent;
 import nomadrealms.context.game.event.Target;
+import nomadrealms.context.game.item.Inventory;
 import nomadrealms.context.game.world.World;
 import nomadrealms.context.game.world.map.area.Tile;
+import nomadrealms.context.game.world.map.area.coordinate.TileCoordinate;
 import nomadrealms.event.game.effect.EffectContext;
 import nomadrealms.render.Renderable;
+import nomadrealms.render.particle.NullParticlePool;
 import nomadrealms.render.particle.ParticlePool;
 import nomadrealms.render.particle.spawner.BasicParticleSpawner;
 
@@ -31,31 +35,64 @@ import nomadrealms.render.particle.spawner.BasicParticleSpawner;
  *
  * @author Lunkle
  */
-public interface Actor extends HasPosition, HasHealth, HasInventory, Target, Renderable {
+@Derializable
+public abstract class Actor implements HasPosition, HasHealth, HasInventory, Target, Renderable {
 
-	UUID uuid();
+	private transient final UUID uuid = UUID.randomUUID();
 
-	String name();
+	private transient ParticlePool particlePool = new NullParticlePool();
 
-	default List<Action> actions() {
+	protected TileCoordinate tileCoord;
+	protected transient Tile tile;
+	protected int health;
+	protected boolean dead;
+	protected final Status status = new Status();
+	protected final Inventory inventory = new Inventory(this);
+	protected String name;
+
+	/**
+	 * No-arg constructor for serialization.
+	 */
+	protected Actor() {
+	}
+
+	protected Actor(String name) {
+		this.name = name;
+	}
+
+	public UUID uuid() {
+		return uuid;
+	}
+
+	public String name() {
+		return name;
+	}
+
+	public List<Action> actions() {
 		return new ArrayList<>();
 	}
 
-	default void update(GameState state) {
+	public void update(GameState state) {
 	}
 
-	default List<InputEvent> retrieveNextPlays() {
+	public List<InputEvent> retrieveNextPlays() {
 		return new ArrayList<>();
 	}
 
-	boolean dead();
+	public boolean dead() {
+		return dead;
+	}
 
-	void dead(boolean dead);
+	public void dead(boolean dead) {
+		this.dead = dead;
+	}
 
-	Status status();
+	public Status status() {
+		return status;
+	}
 
 	@Override
-	default void damage(int damage) {
+	public void damage(int damage) {
 		if (damage > 0 && status().count(INVINCIBLE) > 0) {
 			status().remove(INVINCIBLE, 1);
 			particlePool().addParticles(new SpawnParticlesEffect(
@@ -71,26 +108,32 @@ public interface Actor extends HasPosition, HasHealth, HasInventory, Target, Ren
 		HasHealth.super.damage(damage);
 	}
 
-	void particlePool(ParticlePool particlePool);
+	public void particlePool(ParticlePool particlePool) {
+		this.particlePool = particlePool;
+	}
 
-	ParticlePool particlePool();
+	public ParticlePool particlePool() {
+		return particlePool;
+	}
 
 	/**
 	 * @return Whether this actor has the capability of restocking cards when they run out, or if they will simply die once
 	 * they are out of cards.
 	 */
-	default boolean shouldRestock() {
+	public boolean shouldRestock() {
 		return true;
 	}
 
 	/**
 	 * purely done for the sake of adding references to optimize other algorithms
 	 */
-	default void reindex(World world) {
+	public void reindex(World world) {
+		tile = world.getTile(tileCoord);
+		inventory.reindex(this);
 	}
 
 	@Override
-	default boolean move(Tile target) {
+	public boolean move(Tile target) {
 		if (target.actor() != null) {
 			return false;
 		}
@@ -102,6 +145,32 @@ public interface Actor extends HasPosition, HasHealth, HasInventory, Target, Ren
 		}
 		target.actor(this);
 		return true;
+	}
+
+	@Override
+	public Tile tile() {
+		return tile;
+	}
+
+	@Override
+	public void tile(Tile tile) {
+		this.tile = tile;
+		this.tileCoord = tile == null ? null : tile.coord();
+	}
+
+	@Override
+	public int health() {
+		return health;
+	}
+
+	@Override
+	public void health(int health) {
+		this.health = health;
+	}
+
+	@Override
+	public Inventory inventory() {
+		return inventory;
 	}
 
 }
